refactor: make pippenger_multiproduct_solver generic ( PROOF-641 )
