+++ sxt/multiexp/curve21/BUILD
@@ -6,72 +6,43 @@ load(
 sxt_cc_component(
     name = "multiexponentiation",
     impl_deps = [
-        ":pippenger_multiproduct_solver",
         "//sxt/base/container:blob_array",
-        "//sxt/execution/async:future",
+        "//sxt/base/device:event",
+        "//sxt/base/device:event_utility",
+        "//sxt/base/device:memory_utility",
+        "//sxt/base/device:stream",
+        "//sxt/base/iterator:index_range",
+        "//sxt/curve21/operation:add",
+        "//sxt/curve21/operation:double",
+        "//sxt/curve21/operation:neg",
         "//sxt/curve21/type:element_p3",
+        "//sxt/execution/async:coroutine",
+        "//sxt/execution/async:future",
+        "//sxt/execution/device:device_viewable",
+        "//sxt/execution/device:for_each",
         "//sxt/memory/management:managed_array",
+        "//sxt/memory/resource:async_device_resource",
+        "//sxt/memory/resource:device_resource",
+        "//sxt/multiexp/base:exponent_sequence",
         "//sxt/multiexp/curve:multiexponentiation_cpu_driver",
         "//sxt/multiexp/curve:multiproduct",
         "//sxt/multiexp/curve:multiproducts_combination",
+        "//sxt/multiexp/curve:pippenger_multiproduct_solver",
         "//sxt/multiexp/pippenger:multiexponentiation",
-        "//sxt/base/device:memory_utility",
-        "//sxt/base/device:event",
-        "//sxt/base/device:event_utility",
-        "//sxt/base/iterator:index_range",
-        "//sxt/memory/resource:device_resource",
-        "//sxt/memory/resource:async_device_resource",
         "//sxt/multiexp/pippenger:multiproduct_decomposition_gpu",
-        "//sxt/multiexp/base:exponent_sequence",
-        "//sxt/execution/async:coroutine",
-        "//sxt/execution/device:device_viewable",
-        "//sxt/execution/device:for_each",
-        "//sxt/base/device:stream",
     ],
     test_deps = [
-        "//sxt/memory/management:managed_array",
-        "//sxt/curve21/type:element_p3",
         "//sxt/base/test:unit_test",
-        "//sxt/multiexp/test:multiexponentiation",
-        "//sxt/memory/resource:managed_device_resource",
+        "//sxt/curve21/type:element_p3",
         "//sxt/execution/async:future",
         "//sxt/execution/schedule:scheduler",
+        "//sxt/memory/management:managed_array",
+        "//sxt/memory/resource:managed_device_resource",
+        "//sxt/multiexp/test:multiexponentiation",
     ],
     deps = [
         "//sxt/base/container:span",
         "//sxt/execution/async:future_fwd",
         "//sxt/memory/management:managed_array_fwd",
-    ],
-)
-
-sxt_cc_component(
-    name = "pippenger_multiproduct_solver",
-    impl_deps = [
-        "//sxt/base/container:blob_array",
-        "//sxt/execution/async:future",
-        "//sxt/memory/management:managed_array",
-        "//sxt/multiexp/curve:multiproduct_cpu_driver",
-        "//sxt/multiexp/pippenger_multiprod:multiproduct",
-        "//sxt/multiexp/pippenger_multiprod:active_offset",
-        "//sxt/curve21/type:element_p3",
-        "//sxt/multiexp/index:index_table",
-        "//sxt/multiexp/index:reindex",
-        "//sxt/multiexp/base:generator_utility",
-    ],
-    test_deps = [
-        "//sxt/base/container:blob_array",
-        "//sxt/multiexp/index:index_table",
-        "//sxt/curve21/type:element_p3",
-        "//sxt/curve21/operation:overload",
-        "//sxt/ristretto/type:literal",
-        "//sxt/base/test:unit_test",
-        "//sxt/memory/management:managed_array",
-        "//sxt/execution/async:future",
-    ],
-    deps = [
-        "//sxt/curve21/operation:add",
-        "//sxt/curve21/operation:double",
-        "//sxt/curve21/operation:neg",
-        "//sxt/multiexp/curve:multiproduct_solver",
     ],
 )
